+++ src/main/scala/com/ovoenergy/datastore4s/DatastoreOperation.scala
@@ -3,6 +3,7 @@ package com.ovoenergy.datastore4s
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
 
+/** Representation of an action to be performed by a datastore service */
 final case class DatastoreOperation[+A](op: DatastoreService => Either[DatastoreError, A]) {
 
   def map[B](f: A => B): DatastoreOperation[B] = DatastoreOperation(ds => op(ds).map(f))
@@ -13,6 +14,11 @@ final case class DatastoreOperation[+A](op: DatastoreService => Either[Datastore
 
 }
 
+object DatastoreOperation extends DatastoreErrors {
+  def pure[A](value: A): DatastoreOperation[A] = DatastoreOperation(_ => Right(value))
+  def failure[A](error: String): DatastoreOperation[A] = DatastoreOperation(_ => exception(new RuntimeException(error)))
+}
+
 object DatastoreOperationInterpreter {
 
   def run[A](operation: DatastoreOperation[A])(implicit datastoreService: DatastoreService): Either[DatastoreError, A] =
@@ -20,7 +26,7 @@ object DatastoreOperationInterpreter {
 
   def runF[A](operation: DatastoreOperation[A])(implicit datastoreService: DatastoreService): Try[A] = run(operation) match {
     case Right(a)    => Success(a)
-    case Left(error) => Failure(DatastoreError.asException(error))
+    case Left(error) => Failure(DatastoreError.asThrowable(error))
   }
 
   def runAsync[A](operation: DatastoreOperation[A])(implicit executionContext: ExecutionContext,
@@ -31,6 +37,6 @@ object DatastoreOperationInterpreter {
                                                      datastoreService: DatastoreService): Future[A] =
     runAsync(operation).flatMap {
       case Right(a)    => Future.successful(a)
-      case Left(error) => Future.failed(DatastoreError.asException(error))
+      case Left(error) => Future.failed(DatastoreError.asThrowable(error))
     }
 }